+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Specify absolute path to spring binfile when starting the server
   (#478)
+* Time out after 10 seconds if starting the spring server doesn't work
+  (maybe related to #480, #479)
 
 ## 1.7.0